+++ compiler-utils/src/main/kotlin/com/r0adkll/kimchi/util/ksp/ClassDeclarations.kt
@@ -6,6 +6,7 @@ import com.google.devtools.ksp.getAllSuperTypes
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSName
 import com.squareup.kotlinpoet.ClassName
+import com.squareup.kotlinpoet.ParameterizedTypeName
 import com.squareup.kotlinpoet.ksp.toTypeName
 
 /**
@@ -14,7 +15,10 @@ import com.squareup.kotlinpoet.ksp.toTypeName
  */
 public fun KSClassDeclaration.implements(className: ClassName): Boolean {
   return getAllSuperTypes().any {
-    it.toTypeName() == className
+    when (val typeName = it.toTypeName()) {
+      is ParameterizedTypeName -> typeName.rawType == className
+      else -> typeName == className
+    }
   }
 }